requirements.txt
Flask==1.1.2
imutils==0.5.4
numpy==1.20.3
opencv-python-headless==4.5.5.64
Pillow==9.0.1
scikit_image==0.18.3
waitress==2.1.1
itsdangerous==2.0.1
